# services/lakehouse/openapi/lakehouse.yaml
openapi: 3.0.0
info:
  title: Lakehouse API
  description: |
    Asset storage and retrieval API for document management.
    Search PDFs by content, view files for vision analysis, process with PyMuPDF.

    This is an internal API that agents can discover and use like any external REST API.

    ## Lakehouse as a Compute Cache for Large API Responses

    The Lakehouse is not just document storage. It is a persistence layer for large
    structured blobs that are expensive to re-fetch from source APIs (Gmail, Slack,
    GitHub, Jira, etc.).

    ### The Core Pattern: Fetch → Blob → Ingest → Reuse

    1. **Fetch wide** — use the source API's native query language as a pre-filter.
       Pull raw results (emails, issues, messages, records) in batch using asyncio.gather()
       with a semaphore for concurrency.

    2. **Build a structured blob** — serialize all records into a single .txt or .json
       file using a consistent delimiter format:

       ```
       --- RECORD ---
       ID: {source_id}
       Source: {api_name}
       Date: {timestamp}
       Subject/Title: {label}
       Body:
       {content capped at 600 chars}
       ```

    3. **Ingest once** — POST /api/v1/assets/ingest (base64-encoded). The asset_id
       is your durable handle. Store it. The blob now survives container expiry.

    4. **Reuse across sessions** — any future execute_code() call retrieves the full
       blob via get_asset_text(asset_id) in a single API call, loading it directly
...

...

...

...

...

...

...
servers:
...
paths:
  /api/v1/assets:
    get:
      operationId: list_assets_external
      summary: List assets (external)
      description: List user's assets via JWT authentication
      parameters:
        - name: content_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: cursor
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
          description: Filter by tag value. When omitted, code_script assets are excluded by default.
        - name: query
          in: query
          schema:
            type: string
          description: Fuzzy filename search (case-insensitive contains match).
      responses:
        '200':
          description: Asset list
        '401':
          description: Unauthorized
  /api/v1/assets/search:
    get:
      operationId: search_assets_external
      summary: Search assets (external)
      description: Full-text search via JWT authentication
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: content_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
        '401':
          description: Unauthorized
    post:
      operationId: search_assets_post
      summary: Search assets (JSON body)
      description: |
        POST variant of asset search for agent/MCP callers that prefer
        JSON request bodies over query parameters. Same underlying search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Full-text search query
                content_type:
                  type: string
                  description: Filter by content type
                limit:
                  type: integer
                  default: 20
                  maximum: 100
      responses:
        '200':
          description: Search results
        '400':
          description: Missing required field query
        '401':
          description: Unauthorized
  /api/v1/assets/upload:
    post:
      operationId: upload_asset_external
      summary: Upload asset (multipart)
      description: Upload a file via multipart form data with JWT authentication
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload successful
        '401':
          description: Unauthorized
  /api/v1/assets/session-artifact:
    post:
      operationId: persist_session_artifact
      summary: Upsert a code-execution session artifact
      description: |
        Upsert an asset produced inside a code-execution session. Keyed on
        `(user_id, session_id, artifact_path)` — repeated writes to the same
        path within the same session overwrite rather than accumulate.
        The S3 object is stored at `artifacts/{user_id}/{session_id}/{artifact_path}`.
        Retires the bridge's direct-Mongo `persist_session_artifacts` path.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - session_id
                - artifact_path
              properties:
                file:
                  type: string
                  format: binary
                session_id:
                  type: string
                artifact_path:
                  type: string
                  description: Relative path within the session workspace, e.g. "user_outputs/report.csv"
                artifact_type:
                  type: string
                  description: Optional, e.g. "output", "input"
                source:
                  type: string
                  description: Optional, e.g. "code_execution_output"
                is_ephemeral:
                  type: boolean
                  default: false
                container_id:
                  type: string
                tenant_id:
                  type: string
      responses:
        '200':
          description: Upsert result
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id: { type: string }
                  user_id: { type: string }
                  session_id: { type: string }
                  artifact_path: { type: string }
                  filename: { type: string }
                  content_type: { type: string }
                  size_bytes: { type: integer }
                  checksum_sha256: { type: string }
                  s3_key: { type: string }
                  is_ephemeral: { type: boolean }
                  created_at: { type: string, format: date-time }
                  updated_at: { type: string, format: date-time }
        '400':
          description: Missing user_id, session_id, or artifact_path
        '401':
          description: Unauthorized
        '500':
          description: S3 upload failed
  /api/v1/assets/code-script:
    post:
      operationId: create_code_script_asset
      summary: Store a code-execution script as an asset (with FIFO cap)
      description: |
        Store a Python/script body as an asset tagged `code_script` with
        `source: "code_execution_script"`. Enforces a rolling FIFO cap per user:
        after insertion, any `code_script` assets beyond the newest 5 are
        evicted (both Mongo doc and S3 object). Retires the bridge's direct-Mongo
        `create_code_asset` path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - filename
              properties:
                code:
                  type: string
                  description: The script source code
                filename:
                  type: string
                  description: Display filename for the script (e.g. "session.py")
      responses:
        '200':
          description: Script asset created
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id: { type: string }
                  filename: { type: string }
                  content_type: { type: string }
                  size_bytes: { type: integer }
                  checksum_sha256: { type: string }
                  s3_key: { type: string }
                  tags:
                    type: array
                    items: { type: string }
                  source: { type: string }
                  created_at: { type: string, format: date-time }
                  evicted_asset_ids:
                    type: array
                    items: { type: string }
                    description: asset_ids evicted by the FIFO cap on this write
        '400':
          description: Missing code or filename
        '401':
          description: Unauthorized
  /api/v1/assets/generate-image:
    post:
      operationId: generate_image_asset
      summary: Generate an image from a text prompt (async)
      description: |
        Creates a placeholder asset immediately and kicks off image generation
        (OpenAI gpt-image-2) in the background. Returns 202 with an `asset_id`
        and `status: "generating"`.

        ### Checking status

        Use `GET /api/v1/assets/{asset_id}` to check progress. The
        `processing_status` field transitions through:

        - `"generating"` — image generation is in progress.
        - `"complete"` — done. Use `/view`, `/download`, or `/bytes` to
          retrieve the image.
        - `"failed"` — generation failed. The `processing_error` field
          contains the reason.

        ### Multi-tasking guidance

        Image generation typically takes 15–60 seconds. The assistant should
        treat this as a background operation and continue working on other
        parts of the user's request rather than blocking on a polling loop.
        Come back to check the asset status after completing other pending
        work, or when the user asks about the image. Avoid tight polling —
        a single check after 30–60 seconds is usually sufficient.

        ### Auth

        The caller's JWT must carry `api_settings.openai.api_key`, or the
        server falls back to `WHISPER_OPENAI_API_KEY` for local development.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
                  description: Text description of the image to generate.
                size:
                  type: string
                  enum: ["1024x1024", "1024x1536", "1536x1024"]
                  default: "1024x1024"
                quality:
                  type: string
                  enum: [low, medium, high]
                  default: "high"
                output_format:
                  type: string
                  enum: [png, webp, jpeg]
                  default: "png"
            examples:
              simple:
                summary: Simple image generation
                value:
                  prompt: "A watercolor painting of a mountain lake at sunrise"
              detailed:
                summary: With size, quality, and format options
                value:
                  prompt: "A minimalist line drawing of a cat sitting on a windowsill"
                  size: "1024x1536"
                  quality: "high"
                  output_format: "png"
      responses:
        '202':
          description: Image generation started. Poll the asset for completion.
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id: { type: string }
                  filename: { type: string }
                  status: { type: string, enum: [generating] }
                  message: { type: string }
              example:
                asset_id: "ast_x7k2m9p4q1n3"
                filename: "generated_ast_x7k2m9p4q1n3.png"
                status: "generating"
                message: "Image generation started. Poll GET /api/v1/assets/{asset_id} for status."
        '400':
          description: Missing prompt or no OpenAI API key available.
        '401':
          description: Unauthorized
  /api/v1/assets/session/{session_id}:
    delete:
      operationId: delete_session_ephemerals
      summary: Bulk-delete ephemeral artifacts for a session
      description: |
        Delete every asset where `user_id` is the caller, `session_id` matches,
        and `is_ephemeral: true`. When `keep_outputs=true`, artifacts whose
        `artifact_path` begins with `user_outputs/` or `drafts/` are preserved
        (they represent user-visible deliverables, not transient session state).
        Retires the bridge's direct-Mongo `delete_ephemeral_artifacts_only` path.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
        - name: keep_outputs
          in: query
          schema:
            type: boolean
            default: false
          description: Preserve artifact_paths under user_outputs/ or drafts/
      responses:
        '200':
          description: Bulk-delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted_count: { type: integer }
                  deleted_asset_ids:
                    type: array
                    items: { type: string }
                  keep_outputs: { type: boolean }
        '400':
          description: Missing user_id or session_id
        '401':
          description: Unauthorized
...
  /api/v1/assets/{asset_id}/download:
    get:
      operationId: download_asset_external
      summary: Download raw file bytes
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Raw file bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        User JWT. Required on every authenticated route.
        Carries `sub` (user_id), `email`, and other identity claims.
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Service-to-service admin key (`MCP_ADMIN_KEY`). Optional.
        When present alongside `BearerAuth`, grants admin-level trust on
        top of the user JWT identity.
...
security:
  - BearerAuth: []
  - BearerAuth: []
    AdminKey: []